sanpay/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==43.0.3
Django==5.0.9
django-celery-beat==2.7.0
django-timezone-field==7.0
djangorestframework==3.15.2
gevent==24.10.3
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
kombu==5.4.2
mssql-django==1.5
packaging==24.2
prompt_toolkit==3.0.48
pycparser==2.22
PyJWT==2.9.0
pyodbc==5.1.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
redis==5.2.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
zope.event==5.0
zope.interface==7.1.1
